fix: Rename the test files for CI is happy
